- Added option to turn off TTS - Added minimal age check - Added notification for new people joining when there is an event currently ongoing

+++ presets.py
@@ -77,27 +77,31 @@ time_formats = [
     '%m/%d/%y %H:%M',
 ]
 
-async def playTTS(text, voice_client):
-    tts = gTTS(text)
-    audio_folder = "audio"
-    if not os.path.exists(audio_folder):
-        os.makedirs(audio_folder)
+async def playTTS(text, voice_client, guild_id):
+    cursor, connection = config.setup()
+    cursor.execute("SELECT tts FROM settings WHERE guild_id=%s", (guild_id,))
+    tts_enabled = cursor.fetchone()[0]
+    if tts_enabled:
+        tts = gTTS(text)
+        audio_folder = "audio"
+        if not os.path.exists(audio_folder):
+            os.makedirs(audio_folder)
 
-    # Generate a unique filename based on timestamp and random number
-    timestamp = datetime.datetime.now().strftime("%Y%m%d_%H%M%S")
-    random_number = random.randint(10, 99)
-    file_name = f"audio_{timestamp}_{random_number}.mp3"
-    file_path = os.path.join(audio_folder, file_name)
+        # Generate a unique filename based on timestamp and random number
+        timestamp = datetime.datetime.now().strftime("%Y%m%d_%H%M%S")
+        random_number = random.randint(10, 99)
+        file_name = f"audio_{timestamp}_{random_number}.mp3"
+        file_path = os.path.join(audio_folder, file_name)
 
-    tts.save(file_path)
-    try:
-        voice_client.play(FFmpegPCMAudio(file_path))
-        while voice_client.is_playing():
-            await asyncio.sleep(1)
-    except Exception as e:
-        print(f"An error occurred while playing TTS: {e}")
-    finally:
-        os.remove(file_path)
+        tts.save(file_path)
+        try:
+            voice_client.play(FFmpegPCMAudio(file_path))
+            while voice_client.is_playing():
+                await asyncio.sleep(1)
+        except Exception as e:
+            print(f"An error occurred while playing TTS: {e}")
+        finally:
+            os.remove(file_path)
 
 async def send_http_request(url, payload):
     async with aiohttp.ClientSession() as session:
